crawler/requirements.txt
requests
beautifulsoup4
fastapi
uvicorn[standard]
python-dotenv
requests
supabase
httpx
async-timeout>=4.0
aiohttp
python-dotenv
pydantic
instructor
pydantic-ai

#genai
google-generativeai
anthropic
openai
xai_sdk
groq
huggingface_hub
mistralai

# monitoring
prometheus_client

#movie databases
tqdm
aiolimiter
